Add toolkit-specific Pingee tests (#315)
* Add toolkit-specific Pingee tests

* Remove old pinger tests

* No need to store the listener on the TestCase any more

* Update traits_futures/testing/gui_test_assistant.py
